+++ QuitMenuOptions.js
@@ -1,21 +1,17 @@
-const options = {}
-
-options.Quit = {
-    label: "Quit"
-    ,description: "⌘Q"
-    ,command: 'workbench.action.quit'
+module.exports = {
+    Quit: {
+        label: "Quit"
+        ,description: "⌘Q"
+        ,command: 'workbench.action.quit'
+    }
+    ,CloseWindow: {
+        label: "Close Window"
+        ,description: "⇧⌘W, ⌘W"
+        ,command: 'workbench.action.closeWindow'
+    }
+    ,Cancel: {
+        label: "Cancel"
+        ,description: "ESC"
+        ,command: "search.action.focusActiveEditor"
+    }
 }
-
-options.CloseWindow = {
-    label: "Close Window"
-    ,description: "⇧⌘W, ⌘W"
-    ,command: 'workbench.action.closeWindow'
-}
-
-options.Cancel = {
-    label: "Cancel"
-    ,description: "ESC"
-    ,command: "search.action.focusActiveEditor"
-}
-
-module.exports = options